Fix wrong return type for PageItem's softShadowBlendMode()
git-svn-id: svn://scribus.net/trunk/Scribus@25583 11d20701-8431-0410-a711-e3c959e3b870

+++ scribus/pageitem.h
@@ -753,7 +753,7 @@ public: // Start public functions
 	double softShadowOpacity() const { return m_softShadowOpacity; }
 
 	void setSoftShadowBlendMode(int val);
-	double softShadowBlendMode() const { return m_softShadowBlendMode; }
+	int softShadowBlendMode() const { return m_softShadowBlendMode; }
 
 	void setSoftShadowErasedByObject(bool val);
 	bool softShadowErasedByObject() const { return m_softShadowErasedByObject; }
